Now is possible to add and edit currencies. Some database changes, but is not necessary a migration at this point.

+++ app/models/movimiento_saldo.rb
@@ -5,7 +5,7 @@ class MovimientoSaldo
   belongs_to :recarga
   belongs_to :compra
   
-  field :saldo, type: Integer
+  field :saldo, type: Float
   field :tipo, type: Boolean
   field :descripcion, type: String
   field :motivo, type: String
